move removeDraftOrder method to DraftCMORemovalAction class
